requirements.txt
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.14
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.157.0
requests==2.32.3
python-dotenv==1.0.1
langchain-groq>=0.1.14